+++ common.sh
@@ -1,5 +1,5 @@
 #! /usr/bin/env bash
 
-IMAGE_NAME=auto_omics_new
+IMAGE_NAME=omixai
 IMAGE_TAG=0.12.0
 IMAGE_FULL=${IMAGE_NAME}:${IMAGE_TAG}